Formata números inteiros como resultado das divisões

+++ calculadora/__init__.py
@@ -11,4 +11,4 @@ def multiplicação(x, y):
 
 
 def divisão(x, y):
-    return x / y
+    return int(x / y)
